requirements.txt
async-timeout==4.0.3
asyncpg==0.29.0
autopep8==2.1.0
greenlet==3.0.3
pyaes==1.6.1
pycodestyle==2.11.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
SQLAlchemy==2.0.30
typing_extensions==4.11.0
